refactor(SettingTab): 拆分三个页签Tab，将SettingTab缩小到100行以内，只负责管理页签 #57
